+++ lib/glimmer/tk/root_proxy.rb
@@ -27,6 +27,9 @@ module Glimmer
     #
     # Follows the Proxy Design Pattern
     class RootProxy < WidgetProxy
+      REGEX_GEOMETRY = /[x+-]/
+      DEFAULT_DIMENSION = '190'
+      
       def initialize(*args)
         @tk = ::TkRoot.new
       end
@@ -43,18 +46,107 @@ module Glimmer
         Glimmer::DSL::Engine.add_content(self, Glimmer::DSL::Tk::RootExpression.new, keyword, *args, &block)
       end
       
+      def has_attribute?(attribute, *args)
+        %w[width height x y].include?(attribute.to_s) || super
+      end
+      
       def set_attribute(attribute, *args)
-        if attribute.to_s == 'iconphoto'
+        case attribute.to_s
+        when 'iconphoto'
           args[0..-1] = [image_argument(args)]
           super
+        when 'width'
+          @width = args.first.to_i
+          self.geometry = "#{args.first.to_i}x#{@height || DEFAULT_DIMENSION}#{x_sign}#{abs_x}#{y_sign}#{abs_y}"
+        when 'height'
+          @height = args.first.to_i
+          self.geometry = "#{@width || DEFAULT_DIMENSION}x#{args.first.to_i}#{x_sign}#{abs_x}#{y_sign}#{abs_y}"
+        when 'x'
+          self.geometry = "#{@width || DEFAULT_DIMENSION}x#{@height || DEFAULT_DIMENSION}#{args.first.to_i > 0 ? '+' : '-'}#{args.first.to_i.abs}#{y_sign}#{abs_y}"
+        when 'y'
+          self.geometry = "#{@width || DEFAULT_DIMENSION}x#{@height || DEFAULT_DIMENSION}#{x_sign}#{abs_x}#{args.first.to_i > 0 ? '+' : '-'}#{args.first.to_i.abs}"
         else
           super
         end
       end
-
+      
+      def get_attribute(attribute)
+        attribute = attribute.to_s
+        case attribute
+        when 'width'
+          geometry.split(REGEX_GEOMETRY)[0].to_i
+        when 'height'
+          geometry.split(REGEX_GEOMETRY)[1].to_i
+        when 'x'
+          sign_number(x_sign, geometry.split(REGEX_GEOMETRY)[2].to_i)
+        when 'y'
+          sign_number(y_sign, geometry.split(REGEX_GEOMETRY)[3].to_i)
+        else
+          super
+        end
+      end
+      
+      def width
+        get_attribute(:width)
+      end
+      
+      def height
+        get_attribute(:height)
+      end
+      
+      def x
+        get_attribute(:x)
+      end
+      
+      def y
+        get_attribute(:y)
+      end
+      
+      def width=(value)
+        set_attribute(:width, value)
+      end
+      
+      def height=(value)
+        set_attribute(:height, value)
+      end
+      
+      def x=(value)
+        set_attribute(:x, value)
+      end
+      
+      def y=(value)
+        set_attribute(:y, value)
+      end
+      
       # Starts Tk mainloop
       def start_event_loop
         ::Tk.mainloop
+      end
+      
+      private
+      
+      def sign_number(sign, number)
+        "#{sign}1".to_i * number
+      end
+      
+      def abs_x
+        geometry.split(REGEX_GEOMETRY)[2].to_i
+      end
+      
+      def abs_y
+        geometry.split(REGEX_GEOMETRY)[3].to_i
+      end
+      
+      def x_sign
+        geometry_signs[0]
+      end
+      
+      def y_sign
+        geometry_signs[1]
+      end
+      
+      def geometry_signs
+        geometry.chars.select {|char| char.match(/[+-]/)}
       end
     end
   end